Several changes.
Sanity check that the hb values and the hb units are in agreement.

Added predictive_labels parameter to Rmds that specify the variables to be'
used for prediction.

Plot the time series length distributions.

Added message if there are less data points than what is the min.node.size
parameter of learning random forest.